requirements.txt
requests
termcolor
selenium 
beautifulsoup4
google-genai
streamlit
watchdog
